spring setting

+++ profile-back/src/main/java/profile/back/repository/IntroductionRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import profile.back.domain.Introduction;
 
-public interface IntroductionRepository extends JpaRepository<Introduction, Integer> {
+public interface IntroductionRepository extends JpaRepository<Introduction, Long> {
 
 }
